Move test utils to packages
+ Part 1 of a bigger change - moving test suite to auto generated
snapshots.
+ This is to help switch to babel-7 as almost all tests fail because of
the code styling of the generated output.
