+++ package.json
@@ -6,13 +6,13 @@
     "build": "gulp build",
     "changelog": "lerna-changelog",
     "clean": "lerna clean",
-    "coverage": "jest --coverage --maxWorkers=4",
+    "coverage": "jest --coverage --maxWorkers=4 --no-cache",
     "coverage-ci": "codecov -f coverage/coverage-final.json",
     "fix": "eslint --fix .",
     "lint": "eslint .",
     "publish": "git pull origin master --rebase && npm run build && npm test && lerna publish",
     "publish-internal": "git pull origin master --rebase && npm run build && NPM_DIST_TAG=internal ./node_modules/.bin/lerna publish --skip-git",
-    "test": "jest --maxWorkers=4",
+    "test": "jest --maxWorkers=4 --no-cache",
     "test-ci": "npm run bootstrap && npm run build && npm run coverage",
     "updated": "lerna updated",
     "watch": "gulp watch"
@@ -66,6 +66,7 @@
     "coverageDirectory": "./coverage/"
   },
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "utils/*"
   ]
 }
