+++ packages/babel-plugin-minify-constant-folding/__tests__/constant-folding-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("constant-folding-plugin", () => {
   thePlugin(
