+++ packages/babel-plugin-minify-flip-comparisons/__tests__/flip-comparisons-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("flip-comparisons", () => {
   thePlugin(
