+++ packages/babel-plugin-minify-guarded-expressions/__tests__/guarded-expressions-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("guarded-expressions-plugin", () => {
   thePlugin(
