+++ packages/babel-plugin-minify-infinity/__tests__/infinity-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("minify-infinity", () => {
   thePlugin(
