+++ packages/babel-plugin-minify-numeric-literals/__tests__/numeric-literals-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("numeric-literals", () => {
   thePlugin(
