+++ packages/babel-plugin-transform-inline-environment-variables/__tests__/inline-env-var-test.js
@@ -1,7 +1,7 @@
 jest.autoMockOff();
 
 const plugin = require("../src/index");
-const thePlugin = require("../../../utils/test-transform")(plugin);
+const thePlugin = require("test-transform")(plugin);
 
 describe("inline-env-plugin", () => {
   let prev;
