+++ packages/babel-plugin-transform-member-expression-literals/__tests__/transform-member-expression-literals-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("transform-member-expressions-literals-plugin", () => {
   thePlugin(
