+++ packages/babel-plugin-transform-minify-booleans/__tests__/booleans-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("boolean-plugin", () => {
   thePlugin(
