+++ packages/babel-plugin-transform-property-literals/__tests__/transform-property-literals-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("transform-property-literals-plugin", () => {
   thePlugin(
