+++ packages/babel-plugin-transform-remove-debugger/__tests__/remove-debugger-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("remove-debugger-plugin", () => {
   thePlugin(
