+++ packages/babel-plugin-transform-simplify-comparison-operators/__tests__/strict-equals-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("simplify-comparison-operators-plugin", () => {
   thePlugin(
