+++ packages/babel-plugin-transform-undefined-to-void/__tests__/undefined-test.js
@@ -1,8 +1,6 @@
 jest.autoMockOff();
 
-const thePlugin = require("../../../utils/test-transform")(
-  require("../src/index")
-);
+const thePlugin = require("test-transform")(require("../src/index"));
 
 describe("undefined-plugin", () => {
   thePlugin(
