+++ utils/test-transform/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "test-transform",
+  "version": "0.0.0",
+  "private": true,
+  "main": "test-transform.js"
+}
